requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
numpy==2.1.1
huggingface_hub==0.24.6
transformers==4.44.2
loguru==0.7.2
aiohttp==3.9.5
python-dotenv==1.0.1


